pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adte-azure-sentinel-triage-engine"
version = "0.1.0"
description = "Automated Detection Triage Engine for Microsoft Sentinel"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ADTE Team"},
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "azure-identity>=1.15",
    "azure-monitor-query>=1.3",
    "openai>=1.0",
    "typer>=0.9",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
adte = "adte.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
